Renamed key package to jwk and moved it within the oauth package
