+++ oauth/jwk/rsa.go
@@ -1,4 +1,4 @@
-package key
+package jwk
 
 import (
 	"crypto/rand"
